Ref: moving IOSchema and getFinalEndpointInputSchema() to a dedicated file.
